kube-proxy: refactor to remove OS specific conditionals.
