+++ cmd/kube-proxy/app/server_windows.go
@@ -1,0 +1,178 @@
+// +build windows
+
+/*
+Copyright 2017 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package app
+
+import (
+	"net"
+
+	"k8s.io/apimachinery/pkg/types"
+	utilnet "k8s.io/apimachinery/pkg/util/net"
+	clientv1 "k8s.io/client-go/pkg/api/v1"
+	"k8s.io/client-go/tools/clientcmd"
+	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
+	"k8s.io/client-go/tools/record"
+	"k8s.io/kubernetes/cmd/kube-proxy/app/options"
+	"k8s.io/kubernetes/pkg/api"
+	"k8s.io/kubernetes/pkg/api/v1"
+	clientset "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
+	"k8s.io/kubernetes/pkg/proxy"
+	proxyconfig "k8s.io/kubernetes/pkg/proxy/config"
+	"k8s.io/kubernetes/pkg/proxy/userspace"
+	"k8s.io/kubernetes/pkg/proxy/winuserspace"
+	"k8s.io/kubernetes/pkg/util/configz"
+	"k8s.io/kubernetes/pkg/util/exec"
+	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
+	utilnetsh "k8s.io/kubernetes/pkg/util/netsh"
+	nodeutil "k8s.io/kubernetes/pkg/util/node"
+	"k8s.io/kubernetes/pkg/util/oom"
+	"k8s.io/kubernetes/pkg/util/resourcecontainer"
+
+	"github.com/golang/glog"
+)
+
+// NewProxyServerDefault creates a new ProxyServer object with default parameters.
+func NewProxyServerDefault(config *options.ProxyServerConfig) (*ProxyServer, error) {
+	if c, err := configz.New("componentconfig"); err == nil {
+		c.Set(config.KubeProxyConfiguration)
+	} else {
+		glog.Errorf("unable to register configz: %s", err)
+	}
+
+	var netshInterface utilnetsh.Interface
+	var iptInterface utiliptables.Interface
+
+	// Create a iptables utils.
+	execer := exec.New()
+
+	netshInterface = utilnetsh.New(execer)
+
+	// We omit creation of pretty much everything if we run in cleanup mode
+	if config.CleanupAndExit {
+		return &ProxyServer{
+			Config:       config,
+			IptInterface: iptInterface,
+		}, nil
+	}
+
+	// TODO(vmarmol): Use container config for this.
+	var oomAdjuster *oom.OOMAdjuster
+	if config.OOMScoreAdj != nil {
+		oomAdjuster = oom.NewOOMAdjuster()
+		if err := oomAdjuster.ApplyOOMScoreAdj(0, int(*config.OOMScoreAdj)); err != nil {
+			glog.V(2).Info(err)
+		}
+	}
+
+	if config.ResourceContainer != "" {
+		// Run in its own container.
+		if err := resourcecontainer.RunInResourceContainer(config.ResourceContainer); err != nil {
+			glog.Warningf("Failed to start in resource-only container %q: %v", config.ResourceContainer, err)
+		} else {
+			glog.V(2).Infof("Running in resource-only container %q", config.ResourceContainer)
+		}
+	}
+
+	// Create a Kube Client
+	// define api config source
+	if config.Kubeconfig == "" && config.Master == "" {
+		glog.Warningf("Neither --kubeconfig nor --master was specified.  Using default API client.  This might not work.")
+	}
+	// This creates a client, first loading any specified kubeconfig
+	// file, and then overriding the Master flag, if non-empty.
+	kubeconfig, err := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
+		&clientcmd.ClientConfigLoadingRules{ExplicitPath: config.Kubeconfig},
+		&clientcmd.ConfigOverrides{ClusterInfo: clientcmdapi.Cluster{Server: config.Master}}).ClientConfig()
+	if err != nil {
+		return nil, err
+	}
+
+	kubeconfig.ContentType = config.ContentType
+	// Override kubeconfig qps/burst settings from flags
+	kubeconfig.QPS = config.KubeAPIQPS
+	kubeconfig.Burst = int(config.KubeAPIBurst)
+
+	client, err := clientset.NewForConfig(kubeconfig)
+	if err != nil {
+		glog.Fatalf("Invalid API configuration: %v", err)
+	}
+
+	// Create event recorder
+	hostname := nodeutil.GetHostname(config.HostnameOverride)
+	eventBroadcaster := record.NewBroadcaster()
+	recorder := eventBroadcaster.NewRecorder(api.Scheme, clientv1.EventSource{Component: "kube-proxy", Host: hostname})
+
+	var proxier proxy.ProxyProvider
+	var endpointsHandler proxyconfig.EndpointsConfigHandler
+
+	proxyMode := proxyModeUserspace
+
+	glog.V(0).Info("Using userspace Proxier.")
+	// This is a proxy.LoadBalancer which NewProxier needs but has methods we don't need for
+	// our config.EndpointsConfigHandler.
+	loadBalancer := userspace.NewLoadBalancerRR()
+	// set EndpointsConfigHandler to our loadBalancer
+	endpointsHandler = loadBalancer
+
+	proxierUserspace, err := winuserspace.NewProxier(
+		loadBalancer,
+		net.ParseIP(config.BindAddress),
+		netshInterface,
+		*utilnet.ParsePortRangeOrDie(config.PortRange),
+		// TODO @pires replace below with default values, if applicable
+		config.IPTablesSyncPeriod.Duration,
+		config.UDPIdleTimeout.Duration,
+	)
+	if err != nil {
+		glog.Fatalf("Unable to create proxier: %v", err)
+	}
+	proxier = proxierUserspace
+
+	// Create configs (i.e. Watches for Services and Endpoints)
+	// Note: RegisterHandler() calls need to happen before creation of Sources because sources
+	// only notify on changes, and the initial update (on process start) may be lost if no handlers
+	// are registered yet.
+	serviceConfig := proxyconfig.NewServiceConfig()
+	serviceConfig.RegisterHandler(proxier)
+
+	endpointsConfig := proxyconfig.NewEndpointsConfig()
+	endpointsConfig.RegisterHandler(endpointsHandler)
+
+	proxyconfig.NewSourceAPI(
+		client.Core().RESTClient(),
+		config.ConfigSyncPeriod,
+		serviceConfig.Channel("api"),
+		endpointsConfig.Channel("api"),
+	)
+
+	config.NodeRef = &v1.ObjectReference{
+		Kind:      "Node",
+		Name:      hostname,
+		UID:       types.UID(hostname),
+		Namespace: "",
+	}
+
+	conntracker := realConntracker{}
+
+	return NewProxyServer(client, config, iptInterface, proxier, eventBroadcaster, recorder, conntracker, proxyMode)
+}
+
+// Not implemented on Windows
+func (s *ProxyServer) tuneConnTracker() error {
+	return nil
+}
